+++ nextjs/src/app/api/cards/route.ts
@@ -1,20 +1,7 @@
 "use server";
 import { createServiceRoleClient } from "@/utils/supabase/server";
-import { OpenAI } from "openai";
 import { NextResponse } from "next/server";
-
-async function initClients() {
-    try {
-        const supabase = createServiceRoleClient();
-        const openai = new OpenAI({
-            apiKey: process.env.OPENAI_API_KEY
-        });
-        return { supabase, openai };
-    } catch (error) {
-        console.error("Error initializing clients:", error);
-        throw new Error("Failed to initialize clients");
-    }
-}
+import initClients from "@/lib/initClient";
 
 // retrieve thread history
 async function getThreadHistory(
