+++ nextjs/src/app/api/history/route.ts
@@ -1,20 +1,7 @@
 "use server";
 import { createServiceRoleClient } from "@/utils/supabase/server";
-import { OpenAI } from "openai";
 import { NextResponse } from "next/server";
-
-async function initClients() {
-    try {
-        const supabase = createServiceRoleClient();
-        const openai = new OpenAI({
-            apiKey: process.env.OPENAI_API_KEY
-        });
-        return { supabase, openai };
-    } catch (error) {
-        console.error("Error initializing clients:", error);
-        throw new Error("Failed to initialize clients");
-    }
-}
+import initClients from "@/lib/initClient";
 
 // retrieve assistant from database
 async function getAssistant(supabase: ReturnType<typeof createServiceRoleClient>) {
