+++ nextjs/src/app/api/profile/cards/route.ts
@@ -1,21 +1,8 @@
 "use server";
 import { createServiceRoleClient } from "@/utils/supabase/server";
-import { OpenAI } from "openai";
 import { NextResponse } from "next/server";
 import { CardsDifficultyType } from "@/types";
-
-async function initClients() {
-    try {
-        const supabase = createServiceRoleClient();
-        const openai = new OpenAI({
-            apiKey: process.env.OPENAI_API_KEY
-        });
-        return { supabase, openai };
-    } catch (error) {
-        console.error("Error initializing clients:", error);
-        throw new Error("Failed to initialize clients");
-    }
-}
+import initClients from "@/lib/initClient";
 
 async function updateCards(
     supabase: ReturnType<typeof createServiceRoleClient>,
