+++ nextjs/src/lib/initClient.ts
@@ -1,0 +1,15 @@
+import { createServiceRoleClient } from "@/utils/supabase/server";
+import { OpenAI } from "openai";
+
+export default async function initClients() {
+    try {
+        const supabase = createServiceRoleClient();
+        const openai = new OpenAI({
+            apiKey: process.env.OPENAI_API_KEY
+        });
+        return { supabase, openai };
+    } catch (error) {
+        console.error("Error initializing clients:", error);
+        throw new Error("Failed to initialize clients");
+    }
+}
